+++ plugins/backend/backstage-plugin-aws-auth/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-aws-auth
 
+## 0.4.8
+
+### Patch Changes
+
+- 85620abc: Bump packages to backstage version 1.13.0
+
 ## 0.4.7
 
 ### Patch Changes